pyproject.toml
[project]
name = "agent-remembers-trajectories"
version = "0.1.0"
description = "Temporal-based workflow system that enhances agent tasks by retrieving and replaying successful past trajectories"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.5.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",
    "pydantic>=2.5.3",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
